+++ small_data/methods/xent.py
@@ -20,7 +20,7 @@ class CrossEntropyClassifier(BasicAugmentation):
 			target class labels and returning a single loss value.
         """
         #return nn.CrossEntropyLoss(reduction='mean') # xent is default
-        return nn.MSELoss(reduction='mean').float()
+        return nn.MSELoss(reduction='mean')
 
     def get_optimizer(self, 
         model: nn.Module,
